requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
requests
pyTelegramBotAPI
cryptography

